Make animation robust to being interacted with
That is, in addition to blurring, if the viewer interacts with the
animated math field in any way like clicking on it (which moves the
cursor and would mess up the quadratic formula if the typing animation
just proceeded blindly), pause; and in case they changed it by moving
the cursor or typing in extra math, reset the math field to the state it
was in when the animation was paused. That way no matter what, if the
animation completes, you get the quadratic formula.