Update sponsor tree and description
Used dashed line for the "your logo here" thingy, and add some text
below it, and make room for the text by moving the whole thing and its
line up a bit.

With dashed lines for the "your logo here" thingy, the lines of the tree
would look weird if they were the same, so now they're solid thin lines.