Fix code sample too big on iOS
Before: https://git.io/vVRfm
After:  https://git.io/vVRfL

There's actually a kinda interesting story of thoughtful design here.
As we all know, the designers of Safari on iOS noticed that webpages
normally look like shit if you lay them out in a 320px-wide viewport,
so they chose to lay them out in a zoomed-out 980px-wide [0] viewport,
yielding tiny but quite faithful copies of what webpages look like in a
desktop browser, which earned them much fame and glory.

Of course, while it's great and all to so faithfully reproduce what you
see in a desktop browser, you also want to be able to actually read the
page, so they let you double-tap to zoom in on any block of text on the
page until just that block fills the width of the screen. Problem: what
if, even when zoomed in so the block of text fills up the screen, the
text is too small to be readable? It sucks when zoomed-in enough to be
able to read the text to have to scroll horizontally side-to-side
because each line of text in this block is too wide to fit in the
screen. So, they came up with the ingenious solution of increasing the
font size of every block of text such that when you zoom in so that the
block fills the width of the screen, the text is guaranteed to be
readable. This results in a layout that largely resembles what the
webpage would look like in a desktop browser, but tweaked to be
actually usable.

Which brings us to the <pre> tag on MathQuill's homepage. Safari on iOS
has apparently decided it's unreadable when zoomed-in to fill the width
of the screen, and so went ahead and bumped up the font size. There's
just one problem: `white-space: pre` means that unlike normal blocks of
text, this one doesn't wrap, it just gets wider, to the point of
sticking out past the edge of the viewport. Now, even besides the
undesirable side effect of messing up document width, this is strictly a
bug since it doesn't even accomplish it's own goal of making this text
readable when zoomed-in to fill the width of the screen (since without
line wrapping, increasing the font size can't change the ratio of the
font size to the width of the block of text), but I can't help but
forgive them considering the innovative design that lead to this.

This bug isn't even hard to correct for. They offer web developers fine-
grained control over all these features, including viewport width [1]
(to let you deliberately design your page to *not* look like shit in
320px, thus launching Responsive Design as a discipline) and,
fortunately for us, -webkit-text-size-adjust [2] to override this
bumping-up-font-size feature. (If only the assholes who designed touch
events could've gotten the memo on fine-grained control: #531)

Of course, there are gotchas and browser bugs:
- -webkit-text-size-adjust is set to 100% and not `none` to avoid an old
  bug in Webkit (which affected desktop Chrome and Safari in addition to
  Safari on iOS) where it would block zooming from happening at all
    + http://blog.55minutes.com/2012/04/iphone-text-resizing/
    + https://bugs.webkit.org/show_bug.cgi?id=56543
- -moz-text-size-adjust is set to `none` and not 100% because that's all
  that's supported by Firefox for Android, percentages aren't supported
    + http://dbaron.org/log/20111126-font-inflation
    + https://developer.mozilla.org/en-US/docs/Web/CSS/text-size-adjust
- -ms-text-size-adjust is omitted because:

        Note:
        The –ms-text-size-adjust property is ignored where the viewport
        tag is present.

      https://msdn.microsoft.com/library/windows/apps/ff462082(v=vs.105).aspx#BKMK_AdjustingTextSizewithCustomCSS
- Opera Mobile has no equivalent
- Finally, text-size-adjust is set to `none` because it's the natural
  value for this purpose, it's described in the original Safari on iOS
  docs as being specifically for this purpose, it makes sense for a
  browser to support only it and no percentages (like mobile Gecko).
  Note that no browser, mobile or otherwise, supports this unprefixed
  variant anyway.

[0]: https://developer.apple.com/library/ios/documentation/AppleApplications/Reference/SafariWebContent/UsingtheViewport/UsingtheViewport.html#//apple_ref/doc/uid/TP40006509-SW27
[1]: https://developer.apple.com/library/ios/documentation/AppleApplications/Reference/SafariWebContent/UsingtheViewport/UsingtheViewport.html#//apple_ref/doc/uid/TP40006509-SW26
[2]: https://developer.apple.com/library/ios/documentation/AppleApplications/Reference/SafariWebContent/AdjustingtheTextSize/AdjustingtheTextSize.html#//apple_ref/doc/uid/TP40006510-SW16