Fix smooth-scroll to not break the back button
In the Beginning, "Community" in the navbar on the homepage was a normal
link, and clicking it would take you to the Community section and
highlight the heading, and the browser Back button would take you back
to your scroll position at the top of the page where the navbar was.

Then I added smooth-scrolling so that when you click "Community" in the
navbar it would smoothly scroll down to the Community heading, and
highlight it. Problem: if you then hit the browser Back button, the
heading highlight would vanish but you would stay in place, rather than
the expected behavior of going back to the scroll position at the top of
the page where you clicked on the navbar.

Now I'm finally fixing that by doing history.pushState():
  https://developer.mozilla.org/en-US/docs/Web/API/History_API
(falling back to the old technique if unavailable)

Of course the web platform sucks so we have to contend with:
  https://github.com/whatwg/html/issues/639